More spell formula tweaks
This started when FIQ and qt pointed out problems with the cap of 20 on
the computed "effective Int" factor, that most spellcasters have high
Int (if they're not elves with 20 already) and get very little benefit
from using these boosts. Also that it was weird how it capped at a value
lower than players often achieve with the helm of brilliance.

So at first I took the advice and cut out the cap, instead capping at
Int if Int was above 20. But the problems still largely existed. So this
is a bit of a larger change to the spellcasting formula, in which I
abolished "effective Int" entirely, which is hopefully a nice
simplification that makes it easier to understand.

The robe and its counterpart items the quarterstaff and nothing wand are
brought more back into line with what they do in vanilla: cut down on
armor penalties. (They don't stack with each other; this might be
interesting to explore later but I think it might be rather too generous
to Wizards.)

Meanwhile, wielding a non-nothing wand that matches a spell gives a
direct, rather large boost. (I trimmed the bonus slightly from what it
gave in effective Int because it's no longer capped, but on the whole I
think the mechanic is underused and could deserve a bit of a boost.)

Estimating the effects on spellcasting is difficult; it's possible that
this will overpower spellcasters somewhat. If so, it can always change
again.