Identify magic flute/harp on use; give message for fire/frost horn
Some UnNetHack-inspired quality of life tidbits. With the horns, I was
not completely happy with "You start playing the horn. In what
direction?" and then the possibility of no message at all if the bolt
was fired away from any walls or monsters (very possible if the player
doesn't know what they have). So there's a message for that now.

Meanwhile, the magic flute and harp are unambiguous, so having them
become identified is consistent with similar object identifications in
xNetHack. In UnNetHack this gives a "It must be a foo!" message, but
I currently don't have that set up and don't want to make these two
items weird by having them be the only ones that print that message.