Port the Oily Corpses Patch
The original patch is by L but the code here is taken from UnNetHack.
The motivation for it is more flavor than anything, though it is nice to
add a little gimmick to certain uncommonly eaten corpses.

Tins of these monsters are not slippery to eat, unless the tin
preparation method is one of the greasy ones anyway (perhaps the act of
putting it into a tin gets rid of this).