Turn the Vibrating Square level into a special level
In this new design, the level is changed: now the hero starts from one
of two fixed points and the Square is found in one of three fixed
points, in large obvious rooms. Players will probably recognize the
layout as a dark mirror of the Astral Plane.

The motivation for this change is primarily because the "Hunt for the
Magic Square" minigame, as jonadab calls it, is not fun. Sure, the
square is mappable in 3.6+ via confused gold detection or a crystal
ball, but for those lacking those resources it's still a tedious search,
and for those with those resources it's just another psuedo-required
metagame element you need to remember.

The level is implemented as a lua file that happens to use
des.mazewalk(), making it a true special level rather than a random maze
level with added fixings tacked on. It is very possible that a later
Gehennom redesign will remove random maze levels entirely, leaving this
one as the only true maze level in the game.