Fix some problems with inaccessible areas in the Arc quest jungle
A couple of complaints came up about this -- the jungles could generate
with certain parts completely sealed off by trees, which could spell
doom for a player who lacked the resources to get out of them. The
natural answer to that is to add the inaccessibles flag, but the problem
with just that as a solution is that there were way too many of these
tiny pockets -- it would be a bonanza of teleport scrolls, rings, and
wands for anyone who took the time to chop down the forest.

Instead, I used some selection math to isolate the unreachable areas and
turn them into trees, then added inaccessibles to the filler level as a
fallback. It can still end up placing some escape items in e.g. an
island or cut-off peninsula in a water body, but not an inordinate
amount of them.

Note that the other complaint of "arc quest is sometimes blocked by
trees; it should have a guaranteed axe" is not addressed. Sometimes the
trees block the path and will need to be cleared. Providing a guaranteed
axe makes rather cheapens the whole jungle-maze challenge. I may in the
future decide to add another way of destroying trees (e.g. burning them)
but that isn't a super high priority.