Fix: a few small mold-fermentation bugs
Two grammar issues and a segfault fix where it would try to alter the
cost of the potion after it had been dropped and the pointer was null.
Also, a little TDTTOE enhancement in which the message about the potion
fizzing is different if the fruit juice being dipped into is already
fizzy or effervescent.