Add black mold
A new monster? In xNetHack???

This has been planned for quite a while, and only now gotten around to.
The general idea is adding a member to the F class that actually might
pose a threat to the player rather than being firmly in assists-only
status. This is particularly relevant since F monsters grow on moldy
corpses.

The general details of the monster come from SporkHack's gray fungus,
via EvilHack which imported it fairly directly from there. Differences
include:
- It's named "black mold" rather than "gray fungus" because black mold
  is a real-life thing that can grow in living spaces, making the
  inhabitants sick.
- It's modeled after the other molds. It has similar stats, no active
  attacks, etc.
- It doesn't confer poison resistance when eaten. Instead it makes the
  consumer deathly ill when eaten. (Though this is curable by vomiting
  and is therefore labeled as FoodPois -- it's a little weird that the
  sickness from the passive attack is not curable by vomiting and the
  sickness from eating it is, but that's due to the different ways it's
  making you sick, one by having it in your digestive tract and one by
  inhaling spores.)
- It appears quite late for a F monster, with a difficulty of 8.
- The Brewing Patch applies to it, allowing black mold to be fermented
  in fruit juice to later turn to sickness.

Unlike in EvilHack, monster illness is not implemented and so the black
mold's passive attack does nothing when it hits a monster.