Implement behavior for non-Planes air terrain
This does not add a new terrain type -- it's still air -- but instead
defines behavior for what happens when air terrain exists on a
non-endgame dungeon level. Formerly, it was just walkable like regular
floor.

The idea has been discussed for a while, and EvilHack beat me to
implementing it, of having levels containing air terrain in which
stumbling onto it means falling off a cliff to one's likely demise. As
of this commit, no levels are changed; this commit provides the
behavior only.

It required a number of insertions into long-established bits of NetHack
program flow, which I hope has not broken anything. I ended up wrapping
all the logic for "[player/monster/object] falls into air" into three
different functions, which ended up being a good choice that saved a lot
of code duplication.

Unlike in EvilHack, falling into air is not necessarily an instadeath
and does not necessarily destroy any objects. If there's a level for the
entities to fall to, they'll usually end up there, though they may be
banged up in the process. Also unlike in EvilHack, air is treated the
same as water or lava, with the same safeguards to prevent accidentally
typoing into it. Here in the xnethack shop we *care* about our players,
dang it.

The fall damage amount of 20d12 was originally 20d6 (the maximum amount
in D&D) and then I doubled the die size because let's face it, an
average expected value of 70 damage (possibly halved by half physical)
is not going to threaten most players. I also tacked on some physical
status effect punishments for hitting the ground and surviving.

There are a couple behaviors that don't make much sense on the face of
it, which I have ignored because they appear to be vanilla issues:
- Polymorphing a flying monster over air into something which doesn't
  fly does not drop it. I have raised github issue #576 for this in
  vanilla since it also exists for water and lava.
- Monsters cannot be knocked back into air to fall to their deaths. This
  is consistent with vanilla behavior for water and lava and is because
  of goodpos() limiting where they can be knocked to.