Add Vlad's Chasm level
This is from EvilHack and utilizes the new air terrain. The Vlad's Tower
branch is extended by one level, adding this one beneath the rest of the
tower (which is unchanged). It combines a terrain challenge -- crossing
the chasm -- with the threat of monsters who can fly, setting up what
are sure to be some interesting battles.

One thing this takes the opportunity to do is add more vampires to the
Tower, which always struck me as a little odd - apart from Vlad there
are only a few vampires on the top floor. EvilHack throws a truly
enormous swarm of them at the player on this level (56 V in all), which
I don't think is proportionate to the challenge level of the rest of the
tower. So there are several each of vampires and vampire leaders and a
couple demons, but not too many.

The vampires have to generate asleep in order to prevent them from
swarming the player before they even leave the stairs area. So I took a
few measures to make sure the player can't waltz past them: one vampire
is sleeping on the stairs (and combat with it will make noise to wake up
its neighbors), and there are some shriekers on the middle islands which
may wake the whole level.