Add new artifact: Mirror Brand
This is the start of an effort to shake up the stale artifact pool in
vanilla while not going so far as to change the whole balance of the
game around it, dNetHack or EvilHack style. Specifically,
- any new artifacts should be on average as beneficial as the ones
  already in vanilla, or less beneficial,
- it is worthwhile to discourage the practice of sacrificing until you
  get the specific artifact you want, so increasing the artifact pool
  size is not a bad thing,
- there is no particular reason to maintain vanilla's system where all
  non-quest artifacts are weapons.

I based Mirror Brand sort of on dNetHack's mirrorblades (using the
damage roll of your enemy's weapon against them somehow) and also drew
on the characteristics of the existing Brands in xNetHack (short sword,
unaligned, can be dual-wielded with either other Brand). It was a toss
up on whether to make it a broadsword or not, but ultimately more roles
can get skill, and higher skill, in short sword so it becomes more
usable this way.

Initially I was going to use a mirrorblade-like mechanic of comparing
Mirror Blade's own damage roll to a temporary damage roll of the
opponent's weapon, and taking whichever one was higher. However, testing
showed that it practically never got a higher roll from an enemy's
weapon, especially considering that the player is usually going to
enchant their weapons up. Reflection and being made of shatterproof
glass meant it was certainly nothing to sneeze at, but it was a pretty
mediocre artifact weapon. (Chris, the dNetHack author, commented that
mirrorblades there are more for monster vs player use where the player is
likely going to have a good weapon and the monster is not.)

So I buffed it further by still taking the enemy's weapon damage roll -
but adding that on top as bonus damage. This makes it a pretty nice
choice when used against enemies with weapons, and the nice thing is
that there are also plenty of enemies without weapons too, so it doesn't
seem overpowered. But of course a lot of the enemies with weapons are
the ones you want to take down fast...

It was going to be neutral to begin with (because it has some nice
"bringing balance" ethos) but then I made it unaligned because the other
Brands are and because the neutral artifact pool is already
disproportionately large anyway.