Fix: artifact dual wielded in offhand didn't provide extrinsics
This is a problem EvilHack had to contend with as well since it allowed
dual-wielding any two artifacts. Fortunately that isn't the case here,
and prior to Mirror Brand's introduction the only effect of this bug was
that you wouldn't get fire or cold resistance from Fire/Frost Brand in
your offhand. A simple set_artifact_intrinsic in set_twoweap() fixed
that problem.

With Mirror Brand's introduction, however, a few updates had to be made
elsewhere in artifact-reflection-handling code as well.