Fix: worn armor still had reduced weight in bones files
Reported a little while ago by I think raisse - you would be wearing
some armor at 75% of its usual weight, then die, and the player who came
across the bones would find it still at 75% weight and be able to pick
up and carry it around at that reduced weight. Not incredibly
exploitable, but still obviously a bug.

The bones-pile-making code apparently doesn't call setnotworn() or
anything like that - perhaps because it would disrupt stuff in the end
of the game like the status bar if you suddenly were shown to lose
intrinsics or something. So do a narrow fix there, only correcting the
weight.