Fix: o_material not searching container correctly
Reported in https://github.com/k21971/EvilHack/issues/67, and then K2
traced this back to the initial object materials patch in xNetHack. The
problem is that it was using the wrong obj* variable in a loop,
leading to gold detection not revealing gold items that are inside
containers. (As far as anyone knows it's only used for gold detection).