Make the high boots in the Gnomish Sewer rotted
This was always the intention since I first designed the level, but
initially these were not made rotted because I thought it wasn't
currently possible to specify erosion on an item created via
des.object(). Today I found out that it is, but the method of specifying
oeroded2 is pretty weird in order to have it use only one field.

I suppose a refactor could change it so that the sp_lev object struct
has separate oeroded, oeroded2, and oerodeproof fields and the lua
parser reads all of them separately, but not a high enough priority
right now (and I don't want to needlessly introduce more merge conflicts
with vanilla). The boots are rotted, and that's good enough.