Make random paper bags contain a potion of booze
This affects 1) only randomly generated bags, 2) only paper ones, and 3)
only those which generated an item inside. It replaces the item inside
with a potion of booze. (Well, *one* of the items inside, but bags only
get 0-1 items anyway.)

The reason it is only for generated bags and doesn't work on wished ones
(which only affects wizard mode anyway) is because object material
wishing generates the object including the contents and THEN sets its
material.