Elves can always squeeze between two trees
This is from Fourk, though implemented for elf monsters as well as the
player. Only works when actually in an elven form - a player of some
other race polymorphed into an elf will be able to squeeze; an elvish
player polymorphed into something non-elvish will not (unless carrying
less than 600 weight like normal).

The idea is that elves have an affinity for trees and are practiced at
slipping between them.