Seed some boulder traps with a dead archeologist under the boulder
From UnNetHack, and according to its commit the original idea was from
aaxelb.

The code is mostly identical to UnNetHack's. I left in the placement of
the corpse and gear under the boulder, rather than on the trap trigger
square, because I think it's a little nicer if there's the possibility
of walking right by it, avoiding the trap, and not noticing. Triggering
the trap, or disabling it, will reveal it, of course.

What is not the same as UnNetHack is that I guarantee the fedora (and
the bullwhip if present) to be at least +2, to push this towards being a
nice find and not just more junk on the floor.

The dead archeologist would obviously have been another quester for the
Amulet, but unlike the "fake bones piles" later added to the upper
floors in vanilla, the items are not guaranteed cursed, because I
figured that the chance of someone wanting to use them would go from
slim to none if they were. This can be handwaved as the items hitting
the 20% chance of not getting cursed in bones...