Allow transferring items directly from container to container
This is a key quality-of-life goal, and one of the remaining widespread
sources of tedium for vanilla players that have not yet been fixed here.
There's a lot of inventory management in NetHack, and for most players
that tends to come with maintaining a stash of items and periodically
switching items between that and the container they're carrying on their
person. Prior to this commit the only way to do this was to take the
items into the main inventory, then use the other container to put them
inside, which could be especially aggravating if the player carried
close to 52 items and had to do a few at a time.

This commit alleviates some of that by adding an option to 't'ransfer
items into another container. When it's selected, the player is prompted
to choose what other container and can select other ones in inventory or
on the floor beneath them. Once selected, they get the "what do you want
to take out" prompt like normal, and objects go into the second
container instead of inventory. This has all the normal effects one
might expect, such as blowing up bags of holding.

This was accomplished by some finagling with in_container and
out_container, and unfortunately had to introduce a new global variable.
I wish I didn't have to do that, but the alternative was passing the
destination container around as an extra parameter, and that would have
made a big mess with how in_container and out_container are used as
callbacks.

You can't transfer items into a container whose contents or lock status
are unknown. This is because I contemplated the possibility of
using a container, choosing to transfer items, and then suddenly dealing
with autounlock/trap-triggering/bag-of-tricks effects, and said
"definitely not." It would cause so many bugs. In practice, I expect
that players will not often run into this case, usually transferring
items into containers they already know well.