Adjust Cleaver-peaceful interactions for less wordiness
Feedback I've gotten is "why force the user to answer a mandatory prompt
here". I initially resisted the idea but eventually came around to
trying it.

Now most Cleaver interactions with peacefuls fall into whether you did
or did not press F before attacking. In most cases, the player won't
have, so they'll just attack the single target to avoid striking the
peaceful. When they do, it'll attack in the arc, without confirmation.
(And F can't be used when applying an axe in a direction.) There are no
longer any prompts. Initially I was going to keep a prompt around, but
then figured you don't get prompted when you use F to attack a peaceful
monster normally, so why do it here? I'm aware that this may be going
too far because now there are edge cases where the player can
legitimately expect to see some sort of confirmation prompt and not get
one. I guess we'll see whether it happens.