Fix: #wizidentify listing inventory heading as a menu option
qt discovered this - because wizard mode identify puts a _ menu option
before the inventory heading, it was becoming a selectable line with
'a'. Besides the obvious problem, this would cause a problem if an
unidentified item in inventory had 'a' as its letter.

Zero the any object so that add_menu() ignores it.