Fix: Cleaver's should_cleave could check off map
It is possible for the player to hit a monster with Cleaver such that
spaces in the arc are not on the map. The should_cleave function didn't
take this into account, so it was possible to segfault by looking for a
monster there. Add an isok() check to prevent that.

Thanks to paxed for pointing out the bug and qt for finding a reliable
way to make it segfault.