Minimal implementation of meaningful monster stunning
Stunning a monster currently does very little that is useful to the
player. About the most significant effect is that the player gains a
small to-hit bonus against them. This aims to change that by making it
meaningful, because there are several interesting proposals for effects
that stun a monster, such as shield bashing, that go nowhere if stunning
is more or less useless. Also by bringing it closer in line with player
stunning, which is a pretty big debilitation.

I say this is minimal because it doesn't fiddle at all with the
spaghetti code that is monster movement. All it does is set the "appr"
variable to 0 when a monster is stunned, meaning it will neither home in
on the player nor run away, and allow mfndpos() to select other
monsters' squares when stunned, as when confused.

Note that monsters still unstun 10% of the time and unconfuse only 2% of
the time, so confusing a monster is a better long-term effect for it.