Allow diagonal movement through open double doors
This is another nice little convenience feature. A number of special
levels (usually the quests) have double doors leading into some large
open area or hall - but they still worked like any other doors in that
you couldn't move into or out of them diagonally. This permits that, but
you do have to be orthogonally adjacent to the door you're not moving
into, and that door must also be open. Otherwise it works as before.