Avoid possible uninitialized mwither fields in montraits()
I'm not sure this actually allows any uninitialized values to be in
there, since it looks like it is only possible to come from copying an
existing and presumably valid struct monst. But no reason not to
explicitly uninitialize it.

This also means that when a corpse of a withering monster is revived, or
the statue of a withering monster is petrified, the monster will no
longer be withering, which is fine. (Not that a monster that died while
withering should leave a corpse anyway, but still applies to
petrification.)