Izchak occasionally stocks wand/scroll/spellbook of light
Makes sense for him to stock these if he comes across any. I stole the
probability from tallow candles, because who likes tallow candles
anyway. Had to change the iprobs in struct shclass to accommodate this
many items because it was originally designed for only 6. Since this is
constant data it shouldn't be saved or break save files.