Port modified version of L's Bullwhip Patch
The goal of this patch was to make the bullwhip somewhat less of a
quirky early game Archeologist-only thing. It doesn't fundamentally
change that - most players will still want to ditch it for an artifact
at some point - but it makes it more useful to have around.

Applying a whip computes a "proficiency" score for certain things such
as grabbing a monster's weapon or grabbing an item from the floor - this
modifies that to be more based on skill and less on Dex. Also, it adds a
whip-crack mechanic which can be used to frighten animals.

I added a resistance check which the original patch didn't have; it just
felt wrong to not give them a chance to resist (most animals will have
low monster MR anyway) and increased the radius it can affect animals in
since it's not very useful if it only works once the monster moves
into biting range.