Allow erodeproofing destroyable items to make them indestructible
This was a 7.0 feature that I had been putting off for a long time, but
it doesn't seem to have been so bad. This is another object destruction
nerf, and one done in various ways by a few variants up to this point,
effectively allowing the player to choose pieces of gear they really
don't want destroyed and protect them.

This neatly accomplishes a few things:
- It allows for some interesting strategies with items that would
  normally always be in a bag because the player could get hit by
  destructive effects at any time. For instance, the wielding-spellbooks
  mechanic was unworkable in Gehennom because it might burn up in any
  fire trap. Now you can prevent it from doing that.
- In a normal game, the player accumulates more scrolls of enchant
  weapon than they tend to need. The enchantment patches earlier this
  year helped by adding demand for them. This will do the same and
  broaden the set of items on which enchant weapon can be used.
- Frustration when the player's only ring of free action or teleport
  control or wand of death or whatever blows up is now less likely to
  happen, especially in the late game.

Indestructible scrolls of scare monster are immune to fire and blanking
like all other scrolls, but it does nothing for preventing it from
crumbling to dust. This is intentional.