Fix: polyselfing while wearing dragon armor did not work
It appears I broke this in the initial dtsund-DSM commit, by making
draconic polymorphs require a certain value be passed to polyself
instead of checking what you're wearing when polyself is called
regardless of its argument. As a result, the only time polymorphing
would turn you into a dragon was with confused enchant armor melding
dragon scales; other polymorphs would ignore worn dragon armor
completely and just turn you into something random.

This brings it mostly back in line with vanilla behavior by making
polyselfing into a dragon depend on worn items. A special argument for
polyself() is still needed (if confused melding scales to armor, you
shouldn't be able to turn into things besides that type of dragon even
with polymorph control). Polymorphing with control will first ask you
an arbitrary monster to turn into, then will ask if you want to become a
dragon (this part was apparently added in the dtsund-DSM commit and not
documented, probably so that confused melding with polymorph control
doesn't unconditionally transform you.)

If wearing both dragon-scaled armor and dragon scales, the
transformation into a dragon will be based on the body armor and will
break the scales. If wearing non-scaled body armor and scales, the
transformation will merge you with the scales and break the body armor
(slightly illogical as that is).

Since cloak-slot armor can now be uskin, a sanity check also needed
updating.