Make monster extrinsics protecting items (mostly) symmetric
xNetHack already did this for cold, fire, and shock destruction of
items. This changes it to use vanilla's new system using
u_adtyp_resistance_obj(), and applying it to monsters as well.

This also strips out the 1% chance added in vanilla that extrinsics
don't protect an item, because having that guarantee that your items
will be safe is the whole point. This also means that getting hit with a
disintegration attack with extrinsic disintegration resistance now
completely safeguards your armor and shield from disintegration instead
of 1% chance it still applies.