Fix: wished-for weapons could have random weapon onames
Flaw in the random weapon names system. These names are only supposed to
generate on randomly found items, not on a newly created one you just
wished up. Due to how wishing works it also meant that you could wish
for a -3 weapon and it could still have a good-sounding name because the
name was created based on the initial random spe given by mksobj() which
then got overridden by the wish.