Object materials: add a sanity check for items of invalid material
Bug reported by raisse of a gem that somehow wound up as mineral, not
gemstone. I don't have any way to figure out where it came from, so this
adds a sanity check for an object with an illegal material to the rest
of the object sanity checks.

It will also make it very annoying to wizard-mode wish for an item of an
invalid material. But finding unusual material bugs may be worth it (and
the wizard-mode player can just avoid wishing for items which shouldn't
exist anyway.)