Fix: mines lighting state was no longer recognized
Pre-Lua, I implemented the Mines being always lit above Minetown and
always dark beneath it. I did this by having a special value of 2 passed
through the special level code, which would hit code in mkmap() that
interpreted it as such. However, in the last vanilla merge, lit is now
expected to be a boolean value, so it will still accept 0 and 1 but not
2.

Rather than changing this to accommodate the 2, I moved the code out of
mkmap() into the sp_lev.c code that calls it. It will now interpret
"random" lighting using the mineslight logic, if you are in the Mines.