Prevent ceiling hiders from hiding when there is no ceiling
Fuzzer discovered bug - vanilla makes an assumption that because the
only levels with no ceiling are the Planes of Water and Air, which do
not have ROOM tiles, a ceiling hider can hide if they are on a ROOM
tile, because then they must be on a level with a ceiling. This is not
the case in xNetHack, so explicitly check for the ceiling.