Fix: crash with 'mini maze' themed room
This crash happened when the mini maze room tried to generate near the
right side of the map and was not able to place in the first location it
tried because some other room already existed there. This uncommon set
of circumstances meant that the x and y positions of the room got shoved
to the right edge of the map, making it instantly invalid and trigger a
lua error (which escalates to a crash).

I eventually determined that the problem was that the loops that check
the room's footprint for interference from other rooms use the same x
and y variables that the room's coordinate is stored in. Thus, if this
loop runs, and it concludes that the room won't fit in that spot, it
will come back around and try with x=80, and that will immediately
break.

Latent in vanilla, I think, since there are no themed rooms that have a
manually specified x and y position (necessary for mini maze because it
must fall on odd numbers). All of vanilla's map-based rooms leave x and
y unspecified, which makes it rerandomize x and y when it comes back
around, so this crash never happens. I've provided this diff to the
devteam but haven't yet heard anything back.

This does mean that if the maze room footprint collides with something,
it will inefficiently try that same footprint 99 more times and declare
failure, because it's not able to randomly choose a new x and y that
satisfy mazewalk.