Fix: obj materials: blanked leather spellbook stayed leather
In vanilla, it's fine to just change the otyp of the book to
SPE_BLANK_PAPER and leave it at that, because that makes it paper. Under
object materials, though, it still sees itself as leather and that's a
problem because SPE_BLANK_PAPER can't be leather. (Both here and in
vanilla, we're ignoring how dipping it in water magically transmutes the
book into ordinary paper...)

This applied to whatever spellbooks had "vellum" and "parchment" as
randomized descriptions, which are the two leather ones.