Fix: worm tooth and boomerang had no defined strike mode
Triggered an impossible when trying to call weaphitmsg() with a worm
tooth, and discovered that it doesn't have a strike mode like most
weapons do. It was already enough of a special snowflake item with
having its material be 0 instead of BONE, now this?

Changed it to PIERCE because Dune sandworm artwork I can find depicts
teeth as the conical pointed variety, not the sharp-edged variety, and
because the crysknife is also a PIERCE weapon.

Also, boomerangs didn't have a defined strike mode either. This gives
them WHACK, since that's how a boomerang does its damage. (Note: wielded
boomerangs and bows, which also don't have a defined strike mode, will
never call weaphitmsg anyway, because they hit a different case for
inefficiently using the weapon that prints "bash".)