Fix: bad check for dragon-scaled armor when getting price
In the recent dtsund-DSM related change to make dragon-scaled armor much
more valuable, it used the zero/nonzero value of the "dragonscales"
field to determine if the armor was dragon-scaled. Problem is,
dragonscales is corpsenm and most objects have a corpsenm of -1. So it
was trying to read objects[-1], which it obviously shouldn't do.

Is_dragon_scaled_armor() exists; I don't know why I didn't use that from
the beginning here.