Fix: gems in containers could have stone material
This is the now-infamous "stone gem" bug first reported by raisse, and
then tripped up several times in the fuzzer's new sanity check for
object materials. With the power of reverse debugging I finally tracked
it down to mkbox_cnts, in which any rocks that generate within a
container are transformed into a gem or glass by setting otyp directly.
This did not account for the material field still being MINERAL.

I put the fix where it would apply to *any* box contents and not just
rocks, since this also potentially applied in the rare case where a wand
of cancellation generates inside a bag of holding and gets changed to
something non-cancellation and of a different material. It re-inits
the object material rather than setting to base because containers can
contain some items such as amulets which are eligible for random ones.

This is an object materials bug, since the object materials system broke
a vanilla assumption that changing otyp automatically changes the
material.