Fix: "[mdef]'s armor blocks [mdef]'s attack"
Simple mixup between magr and mdef.

Even now that the parameters are in the correct order, this could cause
problems if mdef == magr, since mon_nam_too's '(him|her|it)self' won't
work nicely with s_suffix (would produce '...himself's attack' etc).  I
wrote some code to turn 'himself' into 'his own', etc, in s_suffix, but
I think the combo of s_suffix and mon_nam_too only happens in this one
spot, and I'm not sure if it's actually possible for a monster to attack
itself in that way, so I didn't include it.  If it turns out that it can
happen, I can dig that back out later, but it might be simpler to
rephrase the pline.  If someone sees 'himself's' while playing it will
be a good indicator to revisit this.